requirements.txt
attrs==20.2.0
Django==2.2.16
django-debug-toolbar==3.1.1
django-environ==0.4.5
djangorestframework==3.12.1
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pathlib2==2.3.5
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
pytest-django==4.1.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
zipp==1.2.0
